feat: 임시 데이터 기반 채팅 구조 설계 2
