+++ src/components/chat/SideBar.tsx
@@ -11,99 +11,22 @@ import LogoIcon from "../../../public/icons/logo.svg";
 import SideBarItem from "./SideBarItem";
 import { useUserInfoStore } from "@/stores/useUserInfoStore";
 import groupChatsByDate from "@/utils/chat/groupChatsByDate";
+import { ChatPreviewItem } from "@/types/chat";
 
 interface SideBarProps {
+  chatList: ChatPreviewItem[];
+  selectedChatId: number | null;
+  setSelectedChatId: React.Dispatch<React.SetStateAction<number | null>>;
   onClose?: () => void;
 }
 
-// 임시
-const dummyData = [
-  {
-    chatRoomId: 1,
-    title: "강남역 / 월세 / 오피스텔 / 1억",
-    lastMatchingMessage: null,
-    lastMessageAt: "2025-06-19T10:15:23.000Z",
-  },
-  {
-    chatRoomId: 2,
-    title: "노원 / 전세 / 아파트 / 5억",
-    lastMatchingMessage:
-      "이수 지역, 가격대, 매물 유형 등을 설정하려면 '필터 설정하기' 버튼을 이용해주세요.",
-    lastMessageAt: "2025-06-18T17:42:10.000Z",
-  },
-  {
-    chatRoomId: 3,
-    title: "건대입구 / 매매 / 빌라 / 4억",
-    lastMatchingMessage: null,
-    lastMessageAt: "2025-06-17T09:00:00.000Z",
-  },
-  {
-    chatRoomId: 4,
-    title: "이수역 / 전세 / 아파트 / 6억",
-    lastMatchingMessage:
-      "이수 지역, 가격대, 매물 유형 등을 설정하려면 '필터 설정하기' 버튼을 이용해주세요.",
-
-    lastMessageAt: "2025-06-15T22:33:46.864Z",
-  },
-  {
-    chatRoomId: 5,
-    title: "홍대 / 월세 / 원룸 / 7천만",
-    lastMatchingMessage: "관공서가 가까운 곳을 추천해줘",
-
-    lastMessageAt: "2025-06-16T13:12:54.000Z",
-  },
-  {
-    chatRoomId: 6,
-    title: "잠실 / 매매 / 아파트 / 10억",
-    lastMatchingMessage:
-      "이수 지역, 가격대, 매물 유형 등을 설정하려면 '필터 설정하기' 버튼을 이용해주세요.",
-
-    lastMessageAt: "2025-06-14T19:45:00.000Z",
-  },
-  {
-    chatRoomId: 7,
-    title: "역삼 / 전세 / 오피스텔 / 3억",
-    lastMatchingMessage:
-      "이수 지역, 가격대, 매물 유형 등을 설정하려면 '필터 설정하기' 버튼을 이용해주세요.",
-
-    lastMessageAt: "2025-06-18T08:25:12.000Z",
-  },
-  {
-    chatRoomId: 8,
-    title: "합정 / 월세 / 투룸 / 1억 5천",
-    lastMatchingMessage:
-      "이수 지역, 가격대, 매물 유형 등을 설정하려면 '필터 설정하기' 버튼을 이용해주세요.",
-
-    lastMessageAt: "2025-06-13T11:05:30.000Z",
-  },
-  {
-    chatRoomId: 9,
-    title: "노원 / 매매 / 아파트 / 6억 2천",
-    lastMatchingMessage:
-      "이수 지역, 가격대, 매물 유형 등을 설정하려면 '필터 설정하기' 버튼을 이용해주세요.",
-
-    lastMessageAt: "2025-06-12T20:50:00.000Z",
-  },
-  {
-    chatRoomId: 10,
-    title: "신림 / 전세 / 빌라 / 2억 5천",
-    lastMatchingMessage:
-      "이수 지역, 가격대, 매물 유형 등을 설정하려면 '필터 설정하기' 버튼을 이용해주세요.",
-
-    lastMessageAt: "2025-06-11T15:40:00.000Z",
-  },
-];
-
-const SideBar = ({ onClose }: SideBarProps) => {
+const SideBar = ({ chatList, selectedChatId, setSelectedChatId, onClose }: SideBarProps) => {
   const [searchKeyword, setSearchKeyword] = useState("");
   const [isFocused, setIsFocused] = useState(false);
-  const [selectedChatId, setSelectedChatId] = useState(1);
-  const [searchResults, setSearchResults] = useState([]); // API 응답 값 저장할 상태
-
-  // TODO: 채팅 목록 조회 API
+  const [searchResults, setSearchResults] = useState<ChatPreviewItem[]>([]); // API 응답 값 저장할 상태
 
   // (임시) 검색어 기준 필터링
-  const filteredChats = dummyData.filter((chat) =>
+  const filteredChats = chatList.filter((chat) =>
     chat.title.toLowerCase().includes(searchKeyword.toLowerCase()),
   );
 
@@ -117,7 +40,7 @@ const SideBar = ({ onClose }: SideBarProps) => {
   const { user } = useUserInfoStore();
 
   const handelNewChat = () => {
-    alert("새로운 대화 시작하기");
+    setSelectedChatId(null);
     onClose?.();
 
     // TODO : 채팅 초기화
