+++ src/types/chat.ts
@@ -3,8 +3,14 @@ export interface ChatPreviewItem {
   chatRoomId: number;
   title: string;
   content?: string;
-  lastMatchingMessage?: string;
+  lastMatchingMessage?: string | null;
   lastMessageAt?: string;
+}
+
+export interface ChatItem {
+  chatRoomId: number;
+  title: string;
+  messages: Message[];
 }
 
 export interface Property {
@@ -20,4 +26,12 @@ export interface Property {
   realEstateTypeName: string;
   netArea: number;
   imageUrl: string | null;
+}
+
+export interface Message {
+  messageId: number;
+  senderType: "CHATBOT" | "USER";
+  content: string;
+  createdAt: string;
+  properties?: Property[];
 }
